requirements.txt
pytest>=7.0.0
pytest-html>=3.2.0
pytest-xdist>=2.5.0
pyyaml>=6.0
jinja2>=3.0
requests>=2.28.0
paramiko>=2.11.0
pandas>=1.5.0
openpyxl>=3.0.0
prometheus-client>=0.14.0
psutil>=5.9.0
